fix lint errors, make tests compile again
