+++ components/movementsensor/gpsnmea/serial_test.go
@@ -7,21 +7,11 @@ import (
 	"go.viam.com/test"
 
 	"go.viam.com/rdk/components/movementsensor"
+	"go.viam.com/rdk/components/movementsensor/rtkutils"
 	"go.viam.com/rdk/logging"
 	"go.viam.com/rdk/resource"
 	rutils "go.viam.com/rdk/utils"
 )
-
-func TestValidateSerial(t *testing.T) {
-	fakecfg := &SerialConfig{}
-	path := "path"
-	err := fakecfg.validateSerial(path)
-	test.That(t, err, test.ShouldBeError, resource.NewConfigValidationFieldRequiredError(path, "serial_path"))
-
-	fakecfg.SerialPath = "some-path"
-	err = fakecfg.validateSerial(path)
-	test.That(t, err, test.ShouldBeNil)
-}
 
 func TestNewSerialMovementSensor(t *testing.T) {
 	var deps resource.Dependencies
@@ -50,7 +40,7 @@ func TestNewSerialMovementSensor(t *testing.T) {
 		API:   movementsensor.API,
 		ConvertedAttributes: &Config{
 			ConnectionType: "serial",
-			SerialConfig: &SerialConfig{
+			SerialConfig: &rtkutils.SerialConfig{
 				SerialPath:     path,
 				SerialBaudRate: 0,
 			},
